% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example nitrate dataset}
\format{
A data frame with 200 rows and 3 variables:
\describe{
\item{time}{Date}
\item{nitrate}{mg/L}
\item{site}{factor: A/B/C}
}
}
\source{
Built in data-raw/example_data.R
}
\usage{
example_data
}
\description{
Example nitrate dataset
}
\keyword{datasets}
