% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrate_hourly}
\alias{nitrate_hourly}
\title{Hourly nitrate by site}
\format{
A tibble/data frame with N rows and 4 variables:
\describe{
\item{date_time}{POSIXct. Timestamp.}
\item{site}{chr. Site identifier.}
\item{sensor}{chr. Sensor identifier or type.}
\item{nitrate_umolL}{dbl. Nitrate concentration (µmol/L).}
}
}
\source{
NEON products (codes ...)
}
\usage{
data(nitrate_hourly)
}
\description{
Hourly nitrate by site
}
\keyword{datasets}
