% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wq_hourly}
\alias{wq_hourly}
\title{Hourly water quality by site}
\format{
A tibble/data frame with N rows and 6 variables:
\describe{
\item{date_time}{POSIXct. Timestamp in local time or UTC (state which one).}
\item{site}{chr. Site identifier (e.g., "XYZ").}
\item{sensor}{chr. Sensor identifier or type.}
\item{cond_uScm}{dbl. Specific conductivity (µS/cm).}
\item{do_mgL}{dbl. Dissolved oxygen (mg/L).}
\item{turb_FNU}{dbl. Turbidity (FNU).}
}
}
\source{
NEON products (list product codes + your processing steps).
}
\usage{
data(wq_hourly)
}
\description{
High-frequency water-quality time series collected from in-situ sensors.
}
\keyword{datasets}
