% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny app}
\usage{
run_app()
}
\description{
This function launches the interactive explorer bundled with the package.
It uses the packaged datasets (no reading from disk).
}
\examples{
\dontrun{
  run_app()
}
}
